formatting, rm warnings, fix tests
