+++ src/wallet/test/mercury.test.js
@@ -10,7 +10,6 @@ import { BTC_ADDR, KEYGEN_SIGN_DATA, FINALIZE_DATA, FUNDING_TXID, SHARED_KEY_ID 
 import * as MOCK_CLIENT from '../mocks/mock_wasm';
 import * as MOCK_SERVER from '../mocks/mock_http_client'
 import { BIP32Interface, BIP32,  fromBase58, fromSeed} from 'bip32';
-import { SwapToken } from '../swap/swap'
 
 let bitcoin = require('bitcoinjs-lib')
 let lodash = require('lodash');
@@ -299,30 +298,6 @@ describe('StateChain Entity', function() {
     });
   });
 });
-
-describe('Swaps', function() {
-  describe('swapTokenSign', function() {
-    test('Gen and Verify', async function() {
-      let proof_key_der = fromSeed(Buffer.from("0123456789abcdef"), bitcoin.networks.bitcoin);
-      console.log("proof_key_der: ", proof_key_der.privateKey.toString("hex"))
-      console.log("proof_key_der: ", Array.from(proof_key_der.privateKey))
-      let pub = proof_key_der.publicKey.toString('hex');
-      console.log("pub: ", pub)
-
-      let swap_token = new SwapToken(
-        "120270b4-1a97-46c8-aed6-6b48bf9ff310",
-        10,
-        10,
-        []);
-
-      let msg = swap_token.to_message()
-      let swap_sig = swap_token.sign(proof_key_der);
-      console.log("swap_sig: ", swap_sig)
-      console.log("swap_sig: ", swap_token.verify_sig(proof_key_der, swap_sig))
-    })
-  });
-});
-
 
 
 const SHARED_KEY = {"public":{"q":{"x":"ccd65d5f8e1b5e36fcc6f75daacfb144fc4b2ef956f5968c0dd871474d38d4dd","y":"c3c7acf1b5084d0ce373a415849c6b3072441b2123cf0102f4f9bede9deed434"},"p2":{"x":"ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477","y":"1d970329ed67b215033fa4f45431aa84d5d8969189b166b8b1c995c71418fc79"},"p1":{"x":"5775571335149beef013e041916f8bd6ad74560f614e4cdcc6285cf2239804a4","y":"5d444f7986a767249ec6f624956169edd31ad34169d1e644b9948bb084155a09"},"paillier_pub":{"n":"18618973404778754190518985428533896252170868175908426298152069620162843266686340593727719917765761117728365267017152884385574431579173532202603789689258042399403850464508838324962482715255812960593375421462405196296118705922516865620487438060730020195057476921619387883344824404749260998392405042001310138908055325868010132968223806254698220349656258811324161297045756648004671649730891558015063255667266103305209518264666426136884754344645705399865342796296732565545547519164996468902145097768529183807656951187924006922284038339794011741856368099322064983876937896831825577803442314444728202710488558740408208235001"},"c_key":[1,[2802809810,2295538597,824023868,2468614206,2382410329,3498283269,542845278,2638857966,882874623,314779007,533459108,4110532496,2066816076,3866382865,1507727392,2622766556,2166819970,745160432,4007494000,2755553791,1211723955,4026434693,34787852,817363057,2111193045,1986153889,989827644,3051061802,308497917,3526483949,1313053198,3239938395,3224233175,2656233176,2735490441,528833738,2992068321,1123445282,1322576068,3634835356,3687088759,804528076,1959677799,1848875502,142865825,2608636890,120148350,3818230595,1121032084,629553958,540686410,3942881735,3193185548,3170665768,3039748817,569567082,3758896658,884071316,2728026343,3841465923,3085691968,80776206,2649739738,2735555777,1101020525,3145673761,1465495093,1214998012,3210306804,2083816658,1336173722,3059272070,4195353313,2583489626,2533717747,2110492144,1090372498,3017243663,1048666077,368350691,3098877136,1984080230,662245415,1960616310,1421703698,1721963016,2678994935,1412068076,1627969329,538142028,2155846472,1586893482,2192750863,4161389416,2390953754,2538228113,3035795711,3182964956,732329291,3763461501,1558634765,1289229671,3058158470,2534690608,3070745896,1836149201,2651117011,3055342589,2032472510,669284634,778163039,3181597927,4052939209,82895660,4122394071,3411990510,165191647,1554192820,2780054189,3864774479,1672712056,3345642895,1192039886,4199617900,2746613170,2010732704,3580851445,234733436]]},"private":{"x2":"5b84bbf6c266c8e45f14290f8dd996445144463426c1a093fe76a163c3b5221f"},"chain_code":[0,[]]};
